Unify the definitions of all secret structs
