+++ generic-ec/src/lib.rs
@@ -208,6 +208,7 @@ pub mod multiscalar;
 mod non_zero;
 mod point;
 mod scalar;
+mod secret;
 mod secret_point;
 mod secret_scalar;
 
@@ -268,8 +269,8 @@ pub use self::{
     non_zero::definition::NonZero,
     point::definition::Point,
     scalar::{Radix16Iter, Scalar},
-    secret_point::definition::SecretPoint,
-    secret_scalar::definition::SecretScalar,
+    secret_point::SecretPoint,
+    secret_scalar::SecretScalar,
 };
 
 /// Curves supported out of the box
