+++ generic-ec/src/secret_point.rs
@@ -1,12 +1,29 @@
 use core::fmt;
 use core::iter::Sum;
 
-use crate::{errors::InvalidPoint, Curve, EncodedSecretPoint, Point};
+use crate::{errors::InvalidPoint, secret, Curve, EncodedSecretPoint, Point};
 use subtle::{Choice, ConstantTimeEq};
 
-use self::definition::SecretPoint;
-
-pub mod definition;
+/// Point representing sensitive information (like a derived secret)
+///
+/// Secret point should be treated with an extra care. You shouldn't do any
+/// branching (e.g. `Eq`, `Ord`) on the secret to avoid timing side-channel
+/// attacks, so it implements only constant time traits (like [`ConstantTimeEq`]).
+///
+/// Also, when `alloc` feature is enabled, we enforce extra measures:
+///
+/// * Secret point leaves no trace in RAM after it's dropped \
+///   Memory is zeroized after use
+/// * All clones of a secret point refer to the same region in the memory \
+///   I.e. there will always be only one instance of the point in the memory
+///   no matter how many clones you make
+///
+/// All these guarantees can be bypassed by calling `.as_ref()` and obtaining
+/// `&Point<E>` that is not protected from timing attacks, leaving traces in
+/// the memory, etc.
+///
+/// [`ConstantTimeEq`]: subtle::ConstantTimeEq
+pub struct SecretPoint<E: Curve>(secret::Secret<Point<E>>);
 
 impl<E: Curve> Point<E> {
     /// Convert this value into a [`SecretPoint`]. You should do this at the end
@@ -18,6 +35,15 @@ impl<E: Curve> Point<E> {
 }
 
 impl<E: Curve> SecretPoint<E> {
+    /// Constructs a new secret point
+    ///
+    /// Takes the original point by mutable reference instead of taking by value
+    /// to avoid leaving copies of the point on stack. Point behind the
+    /// reference will be zeroized after the function has returned.
+    pub fn new(point: &mut Point<E>) -> Self {
+        Self(secret::new(point))
+    }
+
     /// Returns the generator defined in the curve specs
     pub fn generator() -> Self {
         Self::new(&mut Point::generator().into())
@@ -38,6 +64,18 @@ impl<E: Curve> SecretPoint<E> {
     pub fn from_bytes(bytes: &[u8]) -> Result<Self, InvalidPoint> {
         let mut point = Point::from_bytes(bytes)?;
         Ok(Self::new(&mut point))
+    }
+}
+
+impl<E: Curve> AsRef<Point<E>> for SecretPoint<E> {
+    fn as_ref(&self) -> &Point<E> {
+        &self.0
+    }
+}
+
+impl<E: Curve> Clone for SecretPoint<E> {
+    fn clone(&self) -> Self {
+        Self(self.0.clone())
     }
 }
